Correction de toute les pages de tableau, chaque bouton fonctionne désormais, le infos sont bien récupéré de chaque liste dans le formulaire d'update, reste a faire la sauvegarde de l'update et regler le soucis de department/date

+++ src/main/webapp/WEB-INF/pages/listemployeeonly.jsp
@@ -37,8 +37,8 @@
 							<th>Last Name</th>
 							<th>Start Date</th>
 							<th>Title</th>
-<!-- 							<th>Departement</th>--->
-							<!-- <th>Superior</th>  -->
+ 							<th>Departement</th>
+							<th>Superior</th>  
 							<th class="td-actions">Actions</th>
 						</tr>
 					</thead>
@@ -51,14 +51,12 @@
 								<td><c:out value="${listEmployee.lastName}" /></td>
 								<td><c:out value="${listEmployee.startDate}" /></td>
 								<td><c:out value="${listEmployee.title}" /></td>	
-<!-- 								<td></td>
-								<td></td> -->
-								<%-- <td><c:out value="${listEmployee.department}" /></td> --%>	 				
-								<%--  <td><c:out value="${listEmployee.employee.firstName}" /></td>  --%>
+								<td><c:out value="${manager.department.deptId}" /></td>	 				
+								<td><c:out value="${manager.employee.empId}" /></td>   
 							<td class="text-center">
 							 <a class='btn btn-info btn-xs' href="updateemployee/${listEmployee.empId}">
 							 <span class="glyphicon glyphicon-edit"></span> Edit</a>
-							 <a href="deleteemployee/${listEmployee.empId}" class="btn btn-danger btn-xs">
+							 <a href="deleteemployee/${listEmployee.empId}" class="btn btn-danger btn-xs" onclick="return confirm('Are you sure you want to delete?')">
 							 <span class="glyphicon glyphicon-remove"></span> Del</a>
 							</td>
 						</tr>
